refactor: record hook 폴더 이동
