+++ frontend/src/components/record/hooks/useMemberStudyListData.ts
