+++ frontend/src/components/record/hooks/useStudyData.ts
